requirements.txt
prometheus_client>=0.16.0
PyYAML>=6.0
pytest>=7.0.0
flake8>=6.1.0
